+++ test/browser/basic.js
@@ -3,17 +3,9 @@ var test = require('tape')
 var WebTorrent = require('../../')
 var fs = require('fs')
 
-var data = fs.readFileSync('test/browser/data/data2.3MB.base64')
+var data = 'R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7'
 var img = Buffer.from(data, 'base64')
 img.name = 'img.png'
-
-var data2 = fs.readFileSync('test/browser/data/data2.5MB.base64')
-var img2 = Buffer.from(data2, 'base64')
-img2.name = 'img2.png'
-
-var data3 = fs.readFileSync('test/browser/data/data2.7MB.base64')
-var img3 = Buffer.from(data3, 'base64')
-img3.name = 'img3.png'
 
 function verifyImage (t, err, elem) {
   t.error(err)
@@ -120,19 +112,5 @@ test('WebTorrent.WEBRTC_SUPPORT', function (t) {
 
   client.destroy(function (err) {
     t.error(err, 'client destroyed')
-  })
-})
-
-test('seed Multiple Torrents', function (t) {
-  t.plan(3)
-  var client = new WebTorrent({ dht: false, tracker: false })
-  client.seed(img, function (torrent) {
-    t.equal(torrent.infoHash,'fe83c6629dc58cc08b7809256000ad068a167b95')
-  })
-  client.seed(img2, function (torrent) {
-    t.equal(torrent.infoHash, 'fd5c8c050571e1f8c369541ed5560f7696931e3f')
-  })
-  client.seed(img3, function (torrent) {
-    t.equal(torrent.infoHash,'2b12aaf42343ce8dc999cfe36a3da318b22bacf2')
   })
 })
